pyproject.toml
[project]
name = "fans"
version = "0.0.60"
description = "fans656's python library"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "build>=1.2.2.post1",
    "deepmerge>=2.0",
    "httpx>=0.28.1",
    "janus>=2.0.0",
    "msgpack>=1.1.0",
    "peewee>=3.17.8",
    "pytz>=2024.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sse-starlette>=2.2.1",
    "sseclient>=0.0.27",
    "tabulate>=0.9.0",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "duf>=0.0.12",
    "freezegun>=1.5.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-mock>=3.14.0",
    "twine>=6.0.1",
]
